requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
cryptography==42.0.5
dnspython==2.6.1
elastic-transport==8.15.0
elasticsearch==8.15.1
email-validator==2.0.0.post2
fastapi==0.104.1
fastapi-async-sqlalchemy==0.6.1
fastapi-limiter==0.1.6
fastapi-pagination==0.12.14
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.7
makefun==1.15.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
passlib==1.7.4
pycparser==2.22
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
redis==5.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.11.0
urllib3==2.2.3
uvicorn==0.24.0.post1
yarl==1.9.4
